Chore: Fix string format warnings (#779)

+++ src/felix86/hle/ptrace.cpp
@@ -1350,7 +1350,7 @@ i64 sys_ptrace(felix86_ptrace_request op, pid_t pid, void* addr, void* data) {
         return set_regs(tracer_mode32, remote_state, data);
     }
     default: {
-        WARN("Unimplemented operation: %x", op);
+        WARN("Unimplemented operation: %x", (int)op);
         return -EIO;
     }
     }
